requirements.txt
# 核心依赖
python-dotenv==1.0.0
pyyaml==6.0.1

# 数据处理
pandas==2.1.4
numpy==1.26.2
ta-lib==0.4.28

# 交易所接口
ccxt==4.2.0
python-binance==1.0.19

# 数据库
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# 异步编程
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0

# 回测框架
backtrader==1.9.78.123
vectorbt==0.26.0

# 可视化
plotly==5.18.0
matplotlib==3.8.2
streamlit==1.29.0

# 监控和日志
prometheus-client==0.19.0
python-telegram-bot==20.7

# 工具库
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 代码质量
black==23.12.1
flake8==6.1.0
mypy==1.7.1
